+++ src/Config/Parser/SymfonyYamlParser.php
@@ -7,6 +7,6 @@ use Symfony\Component\Yaml\Parser as BaseYamlParser;
 /**
  * @author Kamil Samigullin <kamil@samigullin.info>
  */
-class SymfonyYamlParser extends BaseYamlParser implements Parser
+class SymfonyYamlParser extends BaseYamlParser implements ParserInterface
 {
 }